+++ static_core/libpandafile/file-inl.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021-2024 Huawei Device Co., Ltd.
+ * Copyright (c) 2021-2025 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -33,6 +33,54 @@ inline File::StringData File::GetStringData(EntityId id) const
     strData.data = sp.data();
 
     return strData;
+}
+
+// CC-OFFNXT(G.FUD.06) switch-case
+inline std::string StringDataToString(File::StringData sd)
+{
+    std::string str = std::string(utf::Mutf8AsCString(sd.data));
+    size_t symPos = 0;
+    while (symPos = str.find_first_of("\a\b\f\n\r\t\v\'\?\\", symPos), symPos != std::string::npos) {
+        std::string sym;
+        switch (str[symPos]) {
+            case '\a':
+                sym = R"(\a)";
+                break;
+            case '\b':
+                sym = R"(\b)";
+                break;
+            case '\f':
+                sym = R"(\f)";
+                break;
+            case '\n':
+                sym = R"(\n)";
+                break;
+            case '\r':
+                sym = R"(\r)";
+                break;
+            case '\t':
+                sym = R"(\t)";
+                break;
+            case '\v':
+                sym = R"(\v)";
+                break;
+            case '\'':
+                sym = R"(\')";
+                break;
+            case '\?':
+                sym = R"(\?)";
+                break;
+            case '\\':
+                sym = R"(\\)";
+                break;
+            default:
+                UNREACHABLE();
+        }
+        str = str.replace(symPos, 1, sym);
+        ASSERT(sym.size() == 2U);
+        symPos += 2U;
+    }
+    return str;
 }
 
 }  // namespace ark::panda_file